Code review feedback
gets rid of the ugly status name - client_called_complete. Instead we keep a hidden field to track the fact that the client is done with the activity.
adds infrastructure to support hidden fields (any field starting with _ is assumed to be hidden and is not exposed to the clients)