This logs an error when we fail to schedule a delayed job.
We use enqueue* method at several places and as we discovered yesterday on uat, the schedule method can fail in rare situations when the id assigned to the delayed job is a duplicate (bug in moped driver). This change would log such failures to allow us to recover from them.