Fixes a bug where a decision could complete prematurely
A fire_and_forget non-timer child in open state was not covered by the earlier logic. I think this was the reason why great_plains/export_invoices workflow decisions were completing even when the activities under it had not completed.

I will do an audit tomorrow to catch any such activities. We do not have a lot of fire and forget activities, so this number shouldn't be big.

I am working on another branch to further clean up this logic, but this fix should get out first.